Begin converting from using SimpleITK to Pythonic libraries
At first, I was using strictly SimpleITK functions because the SimpleITK images require an extra step to become a Numpy array. In addition, when converting from Numpy array to images, the image metadata is lost and takes more lines of code to fix. This can be tedious so my goal was to try and use SimpleITK strictly.

However, after further research I have found that there is much MORE functions within Numpy Pythonic libraries than SimpleITK and that SimpleITK is lacking particular functions that are necessary.

The long term goal is to use Numpy arrays for storing the image data and utilizing as many Python libraries as possible. If SimpleITK is required, then a simple wrapper function will be written to convert from Numpy to SimpleITK images, call the function, and convert back to Numpy arrays.

WIP