Update algorithm for abdominal cavity
Remove morphological closing on fat image mask. This made the fat void masks less accurate. This was added to close the gap around the belly button. Instead, the belly button is manually landmarked via the config file and a line is drawn to connect the gap there.

Remove other morphological operations on fat void mask. Originally, this was used to remove small objects but I found a better way to perform this is via remove_small_objects function and specifying the desired threshold area.