Update lung detection
Sort the lung objects by area and centroid to correctly get the left and right lung.

Previously, the algorithm assumed there would be only two objects in the lung mask for left and right lung. However, there are instances of small pieces of lung that are not connected to the actual lung. So, to fix this, the algorithm sorts by area and chooses the two largest objects for the left and right lung.

In the future, it would be nice to get the contour of the left and right lung to get a better approximation of it.