Add and implement defaultmin function
defaultmin returns the minimum of an array or a specified value (default) if the array is empty

This is used when getting the sagittal left and right boundaries from the lungs. In Subject 15 initial, the left lung is split in an upper and lower section due to the water image having issues and the result is that the left boundary is nonexistent.

The short term solution is to set them all to zero if that is the case. However, in the future, an improvement would be to take all of the segments of the left lung and create one binary mask that follows the contour of each individual piece. It would be able to remove sharp concavities somehow as well.