Add DICOM Multi-frame support (#10)
Previously, the scans from Washington University at St. Louis were what I will call "standard DICOM". These scans adhere to the original DICOM format. The DICOM group created what is called "enhanced DICOM" that provides a number of improvements to the DICOM format. Washington University is switching one of their Siemens scanners to using the enhanced DICOM format. 

Starting in November 2018, all scans will be in enhanced DICOM. Scans about a month before this contain both enhanced and standard DICOM to aid in switching over.

One of the largest differences is that enhanced DICOM allows one DICOM file (.dcm) to contain multiple frames (i.e. a 3D image). The standard DICOM format only allowed a 2D image, or one frame to be stored in the a .dcm file. This would result in many separate files to represent one multidimensional volume. A drawback to this is that the scan metadata must be present entirely in each DICOM file, resulting in large DICOM scans.

I created a library that extends the capabilities of [pydicom](https://github.com/pydicom/pydicom) named [pydicomext](https://github.com/addisonElliott/pydicomext). Originally, this was calle dicom2 and it did not have support for Multi-frame scans.

This PR updates to use pydicomext and removes the dicom2 folder. One additional major change made to the code is that the C-order indexing convention is used for the 3D volumes. This means that the 3D volume is indexed as (z, y, x) where the first dimensional is slowest varying and last dimension is fastest varying. This is opposite of Fortran-ordered indexing which may be more intuitive since it indexes using (x, y, z). However, to stay with Numpy convention of using C-ordering, I decided to switch. The pydicomext library also uses C-order indexing.