Switch to Texas Tech University unrectified scans (#11)
Previously, the algorithm was designed to run on rectified scans from Texas Tech University. The rectification process consisted of a MATLAB script written by Jason Hill that would rotate the upper and lower volumes independently to better align. The script calculated the centroid and identified the spinal cord axis for better alignment.

This process proved to be extremely complex and unnecessary. In addition, significant portions of the upper and lower scans were cut off resulting in an entire vertebrae typically missing. This is non-ideal since Amy Givan's thesis is on quantifying the amount of VAT & SCAT on a vertebra-by-vertebra basis. Additionally, the typical verebrae cutoff was L3 or L2 and those are the locations, according to studies, where a significant measure of SCAT/VAT loss is recorded.

A script was written to move the unrectified TTU scans to where the rectified ones are. Thus, the actual loading of the scans is unchanged.

The config file was switched from XML to YAML for cleaner code. XML is quite confusing to write in Python. Updated configure window and segmentation algorithm to utilize the YAML config file instead.

Added automatic upper & lower scan alignment via the origin and spacing metadata in the NiFTi files. The images are aligned as close as possible by shifting the images around.